Task_50. Add helper class to keep method for getting random user.
